modified pull input to work a year at a time

+++ pull_input.py
@@ -1,32 +1,48 @@
 # pull_input.py
 
+import argparse
 import os
 import time
 
 from aocd import get_data
 
 
-def main():
+def main(
+        year):
     input_dir = 'input'
 
     if not os.path.exists(input_dir):
         os.mkdir(input_dir)
 
-    for year in range(2015, 2025):
-        for day in range(1, 26):
-            print(f'pulling {year} {day}')
-            data = get_data(day=day, year=year)
+    for day in range(1, 26):
+        filename = f'{year}_{day:02}'
+        filename = os.path.join(input_dir, filename)
+        
+        if os.path.exists(filename):
+            print(f'skipping {year} {day}. Input file exists')
 
-            filename = f'{year}_{day:02}'
-            filename = os.path.join(input_dir, filename)
+            continue;
 
-            with open(filename, 'w') as f:
-                f.write(data)
+        print(f'pulling {year} {day}')
 
-            time.sleep(1.5)
+        data = get_data(day=day, year=year)
+
+        with open(filename, 'w') as f:
+            f.write(data)
+
+        time.sleep(1.5)
 
         print(f'year {year} complete')
 
 if __name__ == '__main__':
-    main()
+    parser = argparse.ArgumentParser(
+            prog='pull_input.py',
+            description='Pull input for a year in this C++ AoC repo')
+
+    parser.add_argument('year')
+
+    args = parser.parse_args()
+
+    main(args.year)
+
